requirements.txt
# Core dependencies
uv>=0.1.0
protobuf>=4.0.0
grpcio>=1.50.0
cyclonedds-python>=0.10.0
paho-mqtt>=2.0.0
opentelemetry-api>=1.13.0
opentelemetry-sdk>=1.13.0
opentelemetry-exporter-otlp>=1.13.0
cloudevents>=1.9.0
zstandard>=0.19.0
numpy>=1.22.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
flake8>=5.0.0
black>=22.0.0
mypy>=0.950
build>=0.8.0
twine>=4.0.0